Enable pedantic warnings

+++ include/server/asio/ssl_server.inl
@@ -30,10 +30,10 @@ inline SSLServer<TServer, TSession>::SSLServer(std::shared_ptr<Service> service,
     switch (protocol)
     {
         case InternetProtocol::IPv4:
-            _endpoint = asio::ip::tcp::endpoint(asio::ip::tcp::v4(), port);
+            _endpoint = asio::ip::tcp::endpoint(asio::ip::tcp::v4(), (unsigned short)port);
             break;
         case InternetProtocol::IPv6:
-            _endpoint = asio::ip::tcp::endpoint(asio::ip::tcp::v6(), port);
+            _endpoint = asio::ip::tcp::endpoint(asio::ip::tcp::v6(), (unsigned short)port);
             break;
     }
 }
@@ -56,7 +56,7 @@ inline SSLServer<TServer, TSession>::SSLServer(std::shared_ptr<Service> service,
     if (context == nullptr)
         throw CppCommon::ArgumentException("SSL context is invalid!");
 
-    _endpoint = asio::ip::tcp::endpoint(asio::ip::address::from_string(address), port);
+    _endpoint = asio::ip::tcp::endpoint(asio::ip::address::from_string(address), (unsigned short)port);
 }
 
 template <class TServer, class TSession>
